Creds/Crates claim
