+++ src/components/App.js
@@ -8,7 +8,6 @@ function App() {
         <img src={logo} className="App-logo" alt="DystoPunks V2" />
         <p style={{lineHeight: "calc(12px + 2vmin)"}}>
         Welcome to a tech-filled dystopian future <br/> where only <b>2077 Punks</b> survived. <br/><br/>       
-        <b style={{color: '#E54286'}}>Web under maintenance!</b>
         </p>
       </header>
     </div>
