+++ src/components/Footer.js
@@ -1,23 +1,33 @@
 import '../styles/Text.css';
 import '../styles/Global.css';
-import {Container} from '@material-ui/core';
+import {Button, Container} from '@material-ui/core';
+
+const styleClaim = {
+  backgroundColor: '#27D3CB',
+  color: '#12173D',
+  fontSize: 'calc(12px + 2vmin)',
+  fontFamily: 'm5x7',
+  boxShadow: 'none',
+  borderRadius: '0px',
+
+  '&:hover': {
+      backgroundColor: '#78FAE6',
+      color: '#ffffff',
+      boxShadow: 'none',
+  }
+};
 
 function Footer() {
   return (
     <Container fixed style={{maxWidth: "700px"}}>
     <div className="Text">
-        <p>
-        <h2>Difference of DystoPunks V2</h2>
-        <a href="https://opensea.io/collection/dystopianpunks" target="_blank" rel="noopener noreferrer"><b>DystoPunks V1</b></a> was one of the first certified <a href="https://unofficialpunks.com/" target="_blank" rel="noopener noreferrer"><b>UnofficialPunks</b></a>, they are a 100% handmade experimental project, which turned out to be a success, only 100 exist and yet there is a great community around them.    
-        <br/><br/> 
-        <b style={{color: '#E54286'}}>DystoPunks V2</b> were created for various reasons; To take dystopians to the next level and generate further growth of the dysto community, enhance the rarity system with many improved and better distributed futuristic tech accessories, develop a <b>Punkverse</b> where the community can interact through their <b>Punks</b>, and of course reward all those who have supported the project since the beginning.  
-        <br/><br/> 
-        <b style={{color: '#E54286'}}>V2</b> are similar to <b>V1</b> as they maintain the same <b>Cyberpunk</b> concept and the same color palette but several accessories have been added and improved <b style={{color: '#E54286'}}>(60 traits)</b> as well as a new <b>Punk Type</b>: <b style={{color: '#E54286'}}>Robots</b>, however the <b>V1</b> will maintain unique aspects that will not exist in <b style={{color: '#E54286'}}>V2</b> such as certain combinations, colorful beards and some traits.  
-        </p>
-        <p>
-        <a href="https://twitter.com/DystopianPunks" target="_blank" rel="noopener noreferrer"><b>Twitter</b></a> | <a href="https://discord.gg/7kZvXxutTM" target="_blank" rel="noopener noreferrer"><b>Discord</b></a><br/>
-        <b style={{color: '#E54286'}}>© 2021 DystopianPunks</b><br/><br/>
-        This project pays homage to the original <b style={{color: '#E54286'}}>CryptoPunks</b> and is not affiliated with <b style={{color: '#E54286'}}>Larva Labs</b>.<br/><br/>
+    <Button style={styleClaim} variant="contained" size="small" href="https://opensea.io/collection/dystopunks" target="_blank" rel="noopener noreferrer"><b>Buy at OpenSea</b></Button>
+        <p style={{color: '#FFFFFF'}}>
+        <a href="https://discord.gg/DystoPunks" target="_blank" rel="noopener noreferrer"><b>Discord</b></a> | <a href="https://dystopunks.medium.com/" target="_blank" rel="noopener noreferrer"><b>Medium</b></a> | <a href="https://twitter.com/DystoPunks" target="_blank" rel="noopener noreferrer"><b>Twitter</b></a>
+        <br/>
+        <b style={{color: '#E54286'}}>© 2021 DystoPunks</b><br/><br/>
+        Inspired by <b>CyberKongz</b> and <b>CryptoPunks</b>.<br/>
+        This project is not affiliated with <b>Larva Labs</b>.<br/><br/>
         </p>
     </div>
     </Container>
