+++ src/styles/App.css
@@ -9,7 +9,7 @@
 
 .App-header {
   background-color: #12173d;
-  padding-top: 20vh;
+  padding-top: 10vh;
   display: flex;
   flex-direction: column;
   align-items: center;
