--- src/styles/Roadmap.css
@@ -1,26 +1,0 @@
-.Roadmap {
-  text-align: center;
-  text-align-last: justify;
-  line-height: calc(12px + 1.5vmin);
-}
-
-.Roadmap-logo {
-  height: 10vmin;
-  pointer-events: none;
-}
-
-.Roadmap-header {
-  background-color: #ffffff;
-  padding-top: 1vh;
-  margin-bottom: -3vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(12px + 2vmin);
-  color: #12173d;
-}
-
-.Kanban {
-  height: 2vmin;
-}
